added policy to delete (removed function from model)

+++ app/Http/Controllers/PostController.php
@@ -10,7 +10,7 @@ class PostController extends Controller
     public function index()
     {
         //eager loading
-        $posts = Post::with(['user', 'likes'])->paginate(20);
+        $posts = Post::orderBy('created_at', 'desc')->with(['user', 'likes'])->paginate(20);
         return view('posts.index', [
             'posts' => $posts
         ]);
@@ -28,6 +28,7 @@ class PostController extends Controller
     }
     public function destroy(Post $post)
     {
+        $this->authorize('delete', $post); //policy we defined in PostPolicy
         $post->delete();
         return back();
     }
